test-requirements.txt
requests
beautifulsoup4
PyYAML
six
pathlib